Editor: Reuse data query in the post author components
