+++ packages/editor/src/components/post-author/constants.js
@@ -1,6 +1,10 @@
+export const BASE_QUERY = {
+	_fields: 'id,name',
+	context: 'view', // Allows non-admins to perform requests.
+};
+
 export const AUTHORS_QUERY = {
 	who: 'authors',
 	per_page: 50,
-	_fields: 'id,name',
-	context: 'view', // Allows non-admins to perform requests.
+	...BASE_QUERY,
 };
